chore: move ApiKey type to backends, remove AccountTier sqlx::Type
